Move update coordinator to coordinator.py
Calculate the rest of attributes from built-in statistics

Update websockets to specify the number of records

Add storage support again
