+++ requirements.txt
@@ -1,2 +1,3 @@
 e-data == 0.3.9
 python-dateutil == 2.8.2
+numpy == 1.22.1
